+++ src/encoder_formats/fdkaac_format.ml
@@ -69,7 +69,7 @@ type t = {
   bitrate_mode   : bitrate_mode;
   bitrate        : int;
   channels       : int;
-  samplerate     : int;
+  samplerate     : int Lazy.t;
   sbr_mode       : bool;
   transmux       : transmux
 }
@@ -121,6 +121,6 @@ let to_string m =
   Printf.sprintf "%%fdkaac(afterburner=%b,aot=%S,%s,channels=%d,\
                            samplerate=%d,sbr_mode=%b,transmux=%S)"
     m.afterburner (string_of_aot m.aot) br_info m.channels
-    m.samplerate m.sbr_mode (string_of_transmux m.transmux)
+    (Lazy.force m.samplerate) m.sbr_mode (string_of_transmux m.transmux)
 
 let bitrate m = m.bitrate * 1000